Add functionality to make the switch from a temp to a production environment.
This has quite a bit of moving parts. In the URL bar of the browser, a button is added, which calls background.js to open a new tab with the UI. Additionaly, a script is injected in the original page to perform the API operations. The script from the UI page communicates with the page script using messages.