Complete the publication process by also including the duplication step.
This is actually a bit of a hack, as the Duplicator MediaWiki extension doesn't include an API, so we have to call the front-end. This also performs the link rewriting _after_ the move, so for a very brief moment, the live page includes links to the Vprepub environment. This moment should be brief enough not to have any impact.