First (major) part of some significant rework to generalize this workflow from just MedMij:V2020.01 to all informatiestandaarden pages.
This rework has a bit of a snowball effect; as we need finer-grained control
over what happens on which page, the pattern matching in manifest.json is
replaced by a JS based dispatcher. This in turn allows/requires better re-use
of components and this in turn requires re-architecting these parts, and there
you go.

It is not entirely done, but the the main parts are handled by the dispatcher:
* create new issues from live pages
* merge issue pages to prepub pages